Capnproto serialization (#14)
* Initiating FlatBuffers attempt

* Initiating FlatBuffers experiment

* Flatbuffers loading in progress

* code cleaning: goodbye home made SNLDump
Splitting FBS buffer into interface and implementation

* Splitting FBS buffer interface and implementation

* FBS design interface almost complete

* Initiating capnproto schema and refactoring

* CapnP serialization for interfaces in progress

* Removing FlatBuffers after switch to CapnP
Implementation serialization in progress

* BusNet serialization

* Nets connectivity

* Update build.yml

add capnproto to workflow

* Update build.yml

* Update build.yml

* updating workflows

* Fix CapnP path in valgrind run

* Bug fix in instance model reference serialization
Improving coverage

* Serialization error management

* Introducing back SNL Dump Manifest

* write explicit comparison operator

* Adding support for sub Library sub library

* More test and better code coverage

* Enhancing test coverage

* Coverage exclusion

* Enhancing unit tests and coverage

* more unit testing

* Enhance coverage

+++ test/snl/snl/kernel/SNLDBTest.cpp
@@ -1,12 +1,15 @@
 #include "gtest/gtest.h"
 
 #include "SNLUniverse.h"
+#include "SNLException.h"
 using namespace naja::SNL;
 
 class SNLDBTest: public ::testing::Test {
   protected:
     void TearDown() override {
-      SNLUniverse::get()->destroy();
+      if (SNLUniverse::get()) {
+        SNLUniverse::get()->destroy();
+      }
     }
 };
 
@@ -37,4 +40,12 @@ TEST_F(SNLDBTest, test) {
     
   db1->destroy();
   EXPECT_EQ(universe->getDB(1), nullptr);
+}
+
+TEST_F(SNLDBTest, testErrors) {
+  EXPECT_THROW(SNLDB::create(nullptr), SNLException);
+  SNLUniverse::create();
+  auto db = SNLDB::create(SNLUniverse::get());
+  EXPECT_EQ(1, db->getID());
+  EXPECT_THROW(SNLDB::create(SNLUniverse::get(), SNLID::DBID(1)), SNLException);
 }
